requirements.txt
telethon
aiogram
python-dotenv
google-generativeai
